Update api
Add contactControllers.js
Move all logic from router to controllers
